+++ app/feature/feature-home/src/main/kotlin/com/hedvig/android/feature/home/di/HomeModule.kt
@@ -1,6 +1,7 @@
 package com.hedvig.android.feature.home.di
 
 import com.apollographql.apollo3.ApolloClient
+import com.hedvig.android.core.common.ApplicationScope
 import com.hedvig.android.core.demomode.DemoManager
 import com.hedvig.android.data.chat.read.timestamp.ChatLastMessageReadRepository
 import com.hedvig.android.feature.home.home.data.GetHomeDataUseCaseDemo
@@ -10,6 +11,7 @@ import com.hedvig.android.feature.home.home.data.SeenImportantMessagesStorageImp
 import com.hedvig.android.feature.home.home.ui.HomeViewModel
 import com.hedvig.android.featureflags.FeatureManager
 import com.hedvig.android.memberreminders.GetMemberRemindersUseCase
+import com.hedvig.android.notification.badge.data.crosssell.CrossSellCardNotificationBadgeServiceProvider
 import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
 import org.koin.androidx.viewmodel.dsl.viewModel
@@ -43,6 +45,8 @@ val homeModule = module {
       get<GetHomeDataUseCaseProvider>(),
       get<ChatLastMessageReadRepository>(),
       get<SeenImportantMessagesStorage>(),
+      get<CrossSellCardNotificationBadgeServiceProvider>(),
+      get<ApplicationScope>(),
     )
   }
 }
