+++ app/feature/feature-home/src/main/kotlin/com/hedvig/android/feature/home/home/ui/HomeViewModel.kt
@@ -5,12 +5,22 @@ import com.hedvig.android.data.chat.read.timestamp.ChatLastMessageReadRepository
 import com.hedvig.android.feature.home.home.data.GetHomeDataUseCase
 import com.hedvig.android.feature.home.home.data.SeenImportantMessagesStorage
 import com.hedvig.android.molecule.android.MoleculeViewModel
+import com.hedvig.android.notification.badge.data.crosssell.card.CrossSellCardNotificationBadgeService
+import kotlinx.coroutines.CoroutineScope
 
 internal class HomeViewModel(
   getHomeDataUseCaseProvider: Provider<GetHomeDataUseCase>,
   chatLastMessageReadRepository: ChatLastMessageReadRepository,
   seenImportantMessagesStorage: SeenImportantMessagesStorage,
+  crossSellCardNotificationBadgeServiceProvider: Provider<CrossSellCardNotificationBadgeService>,
+  applicationScope: CoroutineScope,
 ) : MoleculeViewModel<HomeEvent, HomeUiState>(
     HomeUiState.Loading,
-    HomePresenter(getHomeDataUseCaseProvider, chatLastMessageReadRepository, seenImportantMessagesStorage),
+    HomePresenter(
+      getHomeDataUseCaseProvider,
+      chatLastMessageReadRepository,
+      seenImportantMessagesStorage,
+      crossSellCardNotificationBadgeServiceProvider,
+      applicationScope,
+    ),
   )
